Address high and medium priority findbugs warnings
git-svn-id: https://svn.apache.org/repos/asf/xmlgraphics/fop/trunk@1681384 13f79535-47bb-0310-9956-ffa450edef68

+++ src/java/org/apache/fop/render/xml/XMLRenderer.java
@@ -202,7 +202,7 @@ public class XMLRenderer extends AbstractXMLRenderer {
                     continue;
                 }
                 Object value = traitEntry.getValue();
-                if (key == Trait.FONT) {
+                if (((Integer) key).intValue() == Trait.FONT) {
                     FontTriplet triplet = (FontTriplet)value;
                     addAttribute("font-name", triplet.getName());
                     addAttribute("font-style", triplet.getStyle());
@@ -245,8 +245,9 @@ public class XMLRenderer extends AbstractXMLRenderer {
                 } else if (clazz.equals(Color.class)) {
                     Color c = (Color)value;
                     addAttribute(name, ColorUtil.colorToString(c));
-                } else if (key == Trait.START_INDENT || key == Trait.END_INDENT) {
-                    if (((Integer)value).intValue() != 0) {
+                } else if (((Integer) key).intValue() == Trait.START_INDENT
+                        || ((Integer) key).intValue() == Trait.END_INDENT) {
+                    if ((Integer) value != 0) {
                         addAttribute(name, value.toString());
                     }
                 } else {
